handlers: rework the export handler class for the web service
Currently the export classes expose an `run` method that directly
write to a file given in the config. To allow using the exporter in a
web service rework the classes to also provide an `export` method that
only produce the data. The `run` method can then be provided by the
base class using the new `export` method.