Add support for importing Indico Event JSONs (#45)
* Add support for importing Indico Event JSONs

Adds a new ImportHandler which handles Indico Event exports (with the
"detail" export var set to "contributions"), along with a full example
with real data from XDC 2021.

Also makes datetime parsers a bit more universal and able to deal with
3-part timestamps (HH:MM or HH:MM:SS).

* Calculate the fallback speaker ID for invalid events

For Indico contributions where no people are set as speakers/organizers,
we need to set a fallback group (something like "Event Organizers"). The
ID must be a positive int > 0, so let's figure out the highest available
ID and make the fallback ID that max + 1.

* Add configurable track to room name mapping

---------

Co-authored-by: Andi Hubel <40266+saerdnaer@users.noreply.github.com>